renamed package from vocabulary to enums
