+++ vocabulary-rest-ws/src/test/java/org/gbif/vocabulary/restws/resources/documentation/DocumentationBaseTest.java
@@ -3,7 +3,7 @@ package org.gbif.vocabulary.restws.resources.documentation;
 import org.gbif.vocabulary.model.Concept;
 import org.gbif.vocabulary.model.Vocabulary;
 import org.gbif.vocabulary.model.search.KeyNameResult;
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
